[FEATURE] Uncompress ZIP archive files
Added method uncompressZipFile to unzip an archive.
The method takes the zip archive file and the destination directory.

The actual implementation just skips directories found in the archive
as well as it ignores relative path of archived files, extracting them
on the passed directory.

Files in destination folder are overwritten from extracted files having
the same name